requirements.txt
environ~=1.0
pyjwt==2.6.0
python-dotenv==1.0.0
oauthlib==3.0.1
requests==2.29.0
psycopg2-binary